// src/types/room.ts
export interface Amenity {
  name: string;
  description: string;
}

export interface RoomResponse {
  roomId: number;
  roomNumber: number;
  roomPhotoUrl: string;
  roomType: string;
  capacityOfAdults: number;
  capacityOfChildren: number;
  roomAmenities: Amenity[];
  price: number;
  availability: boolean;
}